+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 # Next
 
+* `EKSerializer` no longer implies that serializable objects need to conform to `EKMappingProtocol` or `EKManagedMappingProtocol`.
+
 ## [0.21.1](https://github.com/lucasmedeirosleite/EasyMapping/releases/tag/0.21.1)
 
 * Fixed a bug in `EKSerializer` that prevented has one relationships from being properly serialized with non-nested keypaths.